factor out treniformis updating code and add to dishing update
